requirements.txt
keyboard==0.13.5
openai==0.27.0
PyAudio==0.2.13
python-dotenv==1.0.0
